# Improve performance of `Owners` rule (#16563)