--- a/src/rust/engine/fs/src/lib.rs
+++ b/src/rust/engine/fs/src/lib.rs
@@ -36,7 +36,7 @@ pub use crate::directory::{
   DigestTrie, DirectoryDigest, EMPTY_DIGEST_TREE, EMPTY_DIRECTORY_DIGEST,
 };
 pub use crate::glob_matching::{
-  GlobMatching, PathGlob, PreparedPathGlobs, DOUBLE_STAR_GLOB, SINGLE_STAR_GLOB,
+  FilespecMatcher, GlobMatching, PathGlob, PreparedPathGlobs, DOUBLE_STAR_GLOB, SINGLE_STAR_GLOB,
 };
 
 use std::cmp::min;
@@ -292,29 +292,6 @@ impl GitignoreStyleExcludes {
     match self.gitignore.matched_path_or_any_parents(path, is_dir) {
       ::ignore::Match::None | ::ignore::Match::Whitelist(_) => false,
       ::ignore::Match::Ignore(_) => true,
-    }
-  }
-
-  ///
-  /// Find out if a path has any ignore patterns for files/paths in its tree.
-  ///
-  /// Used by the IntermediateGlobbedFilesAndDirectories in snapshot_ops.rs,
-  /// to check if it may optimize the snapshot subset operation on this tree,
-  /// or need to check for excluded files/directories.
-  ///
-  pub fn maybe_is_parent_of_ignored_path(&self, path: &Path) -> bool {
-    match path.to_str() {
-      None => true,
-      Some(s) => {
-        for pattern in self.exclude_patterns().iter() {
-          if pattern.starts_with(s) || s.starts_with(pattern) {
-            // In case the pattern is shorter than path, we are inside a ignored tree, so both
-            // parent and child of ignored paths.
-            return true;
-          }
-        }
-        false
-      }
     }
   }
 }